+++ views/comment/_form.php
@@ -1,16 +1,17 @@
 <?php if (Yii::app()->user->isGuest) {
-?><div class="comment-not-loggedin">
+?><div class="ext-comment-not-loggedin">
 	Sorry, you have to login to leave a comment.
 </div><?php } else { ?>
-<div id="comment-form-ajax" class="form">
+<div id="ext-comment-form-<?php echo $comment->isNewRecord ? 'new' : 'edit-'.$comment->id; ?>" class="form">
 
 <?php $form = $this->beginWidget('CActiveForm', array(
-	'id'=>'comment-form',
+	'id'=>'ext-comment-form',
     'action'=>array('/comment/comment/create'),
 	'enableAjaxValidation'=>false
 )); ?>
 
-	<?php echo $form->error($comment,'userId'); ?>
+	<?php /** @var CActiveForm $form */
+	echo $form->errorSummary($comment); ?>
 
 	<div class="row">
 		<?php echo $form->labelEx($comment,'message'); ?>
@@ -18,27 +19,38 @@
 		<?php echo $form->error($comment,'message'); ?>
 	</div>
 
-	<?php echo $form->hiddenField($comment, 'type'); ?>
-	<?php echo $form->hiddenField($comment, 'key'); ?>
-
 	<div class="row buttons">
 	    <?php if ($comment->isNewRecord) {
+
+			echo $form->hiddenField($comment, 'type');
+			echo $form->hiddenField($comment, 'key');
+
             /* echo CHtml::hiddenField('returnUrl', $this->createUrl(''));}
 		    echo CHtml::submitButton('Save'); */
 			echo CHtml::ajaxSubmitButton('Submit',
                 array('/comment/comment/create'),
 		        array(
-                    'replace'=>'#comment-form-ajax',
+                    'replace'=>'#ext-comment-form-new',
                     'error'=>"function(){
                         $('#Comment_message').css('border-color', 'red');
                         $('#Comment_message').css('background-color', '#fcc');
                     }"
 		        ),
-		        array('id'=>'submit-comment' . (isset($ajaxId) ? $ajaxId : ''))
+		        array('id'=>'ext-comment-submit' . (isset($ajaxId) ? $ajaxId : ''))
 		    );
-			} else {
-		        echo CHtml::submitButton('Save');
-			}
+		} else {
+			echo CHtml::ajaxSubmitButton('Update',
+				array('/comment/comment/update', 'id'=>$comment->id),
+				array(
+					'replace'=>'#ext-comment-form-edit-'.$comment->id,
+					'error'=>"function(){
+						$('#Comment_message').css('border-color', 'red');
+						$('#Comment_message').css('background-color', '#fcc');
+					}"
+		        ),
+		        array('id'=>'ext-comment-submit' . (isset($ajaxId) ? $ajaxId : ''))
+	        );
+		}
 		?>
 	</div>