+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 ### 0.6.0 (to be released)
 
 - added getCommentCount() to CommentableBehavior (cebe)
+- added delete button and action (cebe)
+- added update button and action (cebe)
 
 ### 0.5.0 (2011-12-19)